+++ src/Rebus.Tests/Performance/TestDispatcherIsolatedSagaPerformance.cs
@@ -27,7 +27,9 @@ namespace Rebus.Tests.Performance
                                         new InMemorySubscriptionStorage(),
                                         new TrivialPipelineInspector(),
                                         new DeferredMessageHandlerForTesting(),
-                                        null);
+                                        null,
+                                        new ReplyDispatcherForTesting(),
+                                        new EndpointInterrogatorForTesting());
         }
 
         /// <summary>