+++ src/Rebus/Bus/IInterrogateThisEndpoint.cs
@@ -1,0 +1,11 @@
+﻿namespace Rebus.Bus
+{
+    interface IInterrogateThisEndpoint
+    {
+        bool OneWayClientMode { get; }
+        string InputQueueAddress { get; }
+        int Workers { get; }
+        int AppDomainRebusEndpointId { get; }
+        bool IsBrokered { get; }
+    }
+}