+++ src/Rebus/Bus/ISendReplies.cs
@@ -1,0 +1,7 @@
+namespace Rebus.Bus
+{
+    interface ISendReplies
+    {
+        void Reply(object reply);
+    }
+}