+++ src/Rebus/Messages/EndpointInterrogationRequest.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Rebus.Messages
+{
+    /// <summary>
+    /// Request that can be sent to an endpoint in order to interrogate it
+    /// </summary>
+    [Serializable]
+    public class EndpointInterrogationRequest : IRebusControlMessage
+    {
+    }
+}